New Features and Bug Fixes

+++ database/migrations/2021_10_08_150325_create_pembelian_detail_table.php
@@ -18,7 +18,7 @@ class CreatePembelianDetailTable extends Migration
             $table->foreignID('id_pembelian')->constrained('pembelian')->onDelete('cascade');
             $table->foreignID('id_item')->constrained('items');
             $table->integer('qty');
-            $table->integer('harga');
+            $table->float('harga', 10, 2);
             $table->integer('subtotal');
             $table->timestamps();
         });
